perf: animation apply — indexed lookup, keyframe cache, float32, bezier LUT
MAR-093, MAR-094: animation_us 81 → 27μs (3x improvement).

- O(1) bone_timeline_index replaces O(N) std::find_if per bone per type
- SamplingContext with cached last_keyframe_index (ozz pattern)
- All keyframe data converted from double to float (halves cache footprint)
- Constant/identity track pruning at load time (256→72 timelines, 63% memory reduction)
- Bezier curves pre-computed into 64-entry LUT (replaces Newton+bisection solver)

+++ src/runtime/binary.cpp
@@ -1069,10 +1069,10 @@ AnimationTranslateBounds compute_translate_bounds(const AnimationData& animation
                 continue;
             }
 
-            bounds.min_x = std::min(bounds.min_x, keyframe.x);
-            bounds.min_y = std::min(bounds.min_y, keyframe.y);
-            bounds.max_x = std::max(bounds.max_x, keyframe.x);
-            bounds.max_y = std::max(bounds.max_y, keyframe.y);
+            bounds.min_x = std::min(bounds.min_x, static_cast<double>(keyframe.x));
+            bounds.min_y = std::min(bounds.min_y, static_cast<double>(keyframe.y));
+            bounds.max_x = std::max(bounds.max_x, static_cast<double>(keyframe.x));
+            bounds.max_y = std::max(bounds.max_y, static_cast<double>(keyframe.y));
         }
     }
     return bounds;
